it no longer removes lines of different colors but when i line of a five or more is created and complete with a unique button it removes ONLY 5 of  the side of the line which contains more of the given color
